MLRERAG-RAG/requirements.txt
pydantic>=2.12.3
fastapi[standard]>=0.120.1
SQLAlchemy>=2.0.44
pydantic-settings>=2.11.0
pgvector>=0.3.6
langchain-core>=1.0.3
langchain-huggingface>=1.0.1
langchain-xai>=1.0.0
langchain_postgres>=0.0.16
PyYAML>=6.0.3
langgraph>=1.0.2
langchain-experimental>=0.3.4
llama-cloud-services>=0.6.77
nest-asyncio>=1.6.0
yake>=0.6.0
arxiv>=2.3.1
psycopg2>=2.9.11
psycopg>=3.2.12
typing_extensions>=4.15.0